Update patch set 7
Patch Set 7:

And no, never add vendor on dependencies and we will never use TheMuppets, it is not our vendor repo.

I will finish this and merge when I see it is good to be merged, until this moment, if you want to suggest ideas, please do it via comments, but don´t touch any part of the commit, please.

Patch-set: 7